同步最新

+++ src/main/java/kohgylw/kiftd/server/util/FileBlockUtil.java
@@ -8,6 +8,10 @@ import kohgylw.kiftd.server.mapper.*;
 import javax.annotation.*;
 import org.springframework.web.multipart.*;
 import java.io.*;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
 import kohgylw.kiftd.server.model.*;
 import kohgylw.kiftd.server.pojo.ExtendStores;
 
@@ -35,7 +39,36 @@ public class FileBlockUtil {
 	@Resource
 	private LogUtil lu;// 日志工具
 	@Resource
-	private FolderUtil fu;
+	private FolderUtil fu;// 文件夹操作工具
+
+	/**
+	 * 
+	 * <h2>清理临时文件夹</h2>
+	 * <p>
+	 * 该方法用于清理临时文件夹（如果临时文件夹不存在，则创建它），避免运行时产生的临时文件堆积。该方法应在服务器启动时和关闭过程中调用。
+	 * </p>
+	 * 
+	 * @author 青阳龙野(kohgylw)
+	 */
+	public void initTempDir() {
+		final String tfPath = ConfigureReader.instance().getTemporaryfilePath();
+		final File f = new File(tfPath);
+		if (f.isDirectory()) {
+			try {
+				Iterator<Path> listFiles = Files.newDirectoryStream(f.toPath()).iterator();
+				while (listFiles.hasNext()) {
+					listFiles.next().toFile().delete();
+				}
+			} catch (IOException e) {
+				lu.writeException(e);
+				Printer.instance.print("错误：临时文件清理失败，请手动清理" + f.getAbsolutePath() + "文件夹内的临时文件。");
+			}
+		} else {
+			if (!f.mkdir()) {
+				Printer.instance.print("错误：无法创建临时文件夹" + f.getAbsolutePath() + "，请检查主文件系统存储路径是否可用。");
+			}
+		}
+	}
 
 	/**
 	 * 
@@ -48,48 +81,89 @@ public class FileBlockUtil {
 	 * @author 青阳龙野(kohgylw)
 	 * @param f
 	 *            MultipartFile 上传文件对象
-	 * @return String 随机生成的保存路径，如果保存失败则返回“ERROR”
+	 * @return java.io.File 生成的文件块，如果保存失败则返回null
 	 */
-	public String saveToFileBlocks(final MultipartFile f) {
-		// 如果存在扩展存储区，则优先在最大的扩展存储区中存放文件（避免占用主文件系统）
+	public File saveToFileBlocks(final MultipartFile f) {
+		// 如果存在扩展存储区，则优先在已有文件块数目最少的扩展存储区中存放文件（避免占用主文件系统）
 		List<ExtendStores> ess = ConfigureReader.instance().getExtendStores();// 得到全部扩展存储区
-		if (ess.size() > 0) {// 如果存在
-			// 找到剩余容量最大的一个
-			ExtendStores maxExtendStores = Collections.max(ConfigureReader.instance().getExtendStores(),
-					new Comparator<ExtendStores>() {
-						@Override
-						public int compare(ExtendStores o1, ExtendStores o2) {
-							// TODO 自动生成的方法存根
-							return (int) (o1.getPath().getFreeSpace() - o2.getPath().getFreeSpace());
+		if (ess.size() > 0) {
+			// 将所有扩展存储区按照已存储文件块的数目从小到大进行排序
+			Collections.sort(ess, new Comparator<ExtendStores>() {
+				@Override
+				public int compare(ExtendStores o1, ExtendStores o2) {
+					try {
+						// 通常情况下，直接比较子文件列表长度即可
+						return o1.getPath().list().length - o2.getPath().list().length;
+					} catch (Exception e) {
+						try {
+							// 如果文件太多以至于超出数组上限，则换用如下统计方法
+							long dValue = Files.list(o1.getPath().toPath()).count()
+									- Files.list(o2.getPath().toPath()).count();
+							return dValue > 0L ? 1 : dValue == 0 ? 0 : -1;
+						} catch (IOException e1) {
+							return 0;
 						}
-					});
-			// 如果该存储区的空余容量大于要存放的文件
-			if (maxExtendStores.getPath().getFreeSpace() > f.getSize()) {
-				final String id = UUID.randomUUID().toString().replace("-", "");
-				final String path = maxExtendStores.getIndex() + "_" + id + ".block";
-				final File file = new File(maxExtendStores.getPath(), path);
-				try {
-					f.transferTo(file);// 则执行存放，并将文件命名为“{存储区编号}_{UUID}.block”的形式
-					return path;
-				} catch (Exception e) {
-					lu.writeException(e);
-					Printer.instance.print(e.getMessage());
+					}
+				}
+			});
+			// 排序完毕后，从文件块最少的开始遍历这些扩展存储区，并尝试将新文件存入一个容量足够的扩展存储区中
+			for (ExtendStores es : ess) {
+				// 如果该存储区的空余容量大于要存放的文件
+				if (es.getPath().getFreeSpace() > f.getSize()) {
+					try {
+						File file = createNewBlock(es.getIndex() + "_", es.getPath());
+						if (file != null) {
+							f.transferTo(file);// 则执行存放，并将文件命名为“{存储区编号}_{UUID}.block”的形式
+							return file;
+						} else {
+							continue;// 如果本处无法生成新的文件块，那么在其他路径下继续尝试
+						}
+					} catch (IOException e) {
+						// 如果无法存入（由于体积过大或其他问题），那么继续尝试其他扩展存储区
+						continue;
+					} catch (Exception e) {
+						lu.writeException(e);
+						Printer.instance.print(e.getMessage());
+						continue;
+					}
 				}
 			}
 		}
 		// 如果不存在扩展存储区或者最大的扩展存储区无法存放目标文件，则尝试将其存放至主文件系统路径下
-		final String fileBlocks = ConfigureReader.instance().getFileBlockPath();
-		final String id = UUID.randomUUID().toString().replace("-", "");
-		final String path = "file_" + id + ".block";
-		final File file = new File(fileBlocks, path);
 		try {
-			f.transferTo(file);// 执行存放，并肩文件命名为“file_{UUID}.block”的形式
-			return path;
+			final File file = createNewBlock("file_", new File(ConfigureReader.instance().getFileBlockPath()));
+			if (file != null) {
+				f.transferTo(file);// 执行存放，并肩文件命名为“file_{UUID}.block”的形式
+				return file;
+			}
 		} catch (Exception e) {
 			lu.writeException(e);
-			Printer.instance.print(e.getMessage());
-			return "ERROR";
+			Printer.instance.print("错误：文件块生成失败，无法存入新的文件数据。详细信息：" + e.getMessage());
 		}
+		return null;
+	}
+
+	// 生成创建一个在指定路径下名称（编号）绝对不重复的新文件块
+	private File createNewBlock(String prefix, File parent) throws IOException {
+		int appendIndex = 0;
+		int retryNum = 0;
+		String newName = prefix + UUID.randomUUID().toString().replace("-", "");
+		File newBlock = new File(parent, newName + ".block");
+		while (!newBlock.createNewFile()) {
+			if (appendIndex >= 0 && appendIndex < Integer.MAX_VALUE) {
+				newBlock = new File(parent, newName + "_" + appendIndex + ".block");
+				appendIndex++;
+			} else {
+				if (retryNum >= 5) {
+					return null;
+				} else {
+					newName = prefix + UUID.randomUUID().toString().replace("-", "");
+					newBlock = new File(parent, newName + ".block");
+					retryNum++;
+				}
+			}
+		}
+		return newBlock;
 	}
 
 	/**
@@ -114,7 +188,8 @@ public class FileBlockUtil {
 	 * 
 	 * <h2>删除文件系统中的一个文件块</h2>
 	 * <p>
-	 * 根据传入的文件节点对象，删除其在文件系统中保存的对应文件块。
+	 * 根据传入的文件节点对象，删除其在文件系统中保存的对应文件块。仅当传入文件节点所对应的文件块不再有其他节点引用时
+	 * 才会真的进行删除操作，否则直接返回true。
 	 * </p>
 	 * 
 	 * @author 青阳龙野(kohgylw)
@@ -123,12 +198,22 @@ public class FileBlockUtil {
 	 * @return boolean 删除结果，true为成功
 	 */
 	public boolean deleteFromFileBlocks(Node f) {
-		// 获取对应的文件块对象
-		File file = getFileFromBlocks(f);
-		if (file != null) {
-			return file.delete();// 执行删除操作
+		// 检查是否还有其他节点引用相同的文件块
+		Map<String, String> map = new HashMap<>();
+		map.put("path", f.getFilePath());
+		map.put("fileId", f.getFileId());
+		List<Node> nodes = fm.queryByPathExcludeById(map);
+		if (nodes == null || nodes.isEmpty()) {
+			// 如果已经无任何节点再引用此文件块，则删除它
+			File file = getFileFromBlocks(f);// 获取对应的文件块对象
+			if (file != null) {
+				return file.delete();// 执行删除操作
+			}
+			return false;
+		} else {
+			// 如果还有，那么直接返回true即可，认为此节点的文件块已经删除了（其他的引用是属于其他节点的）
+			return true;
 		}
-		return false;
 	}
 
 	/**
@@ -161,7 +246,7 @@ public class FileBlockUtil {
 			}
 		} catch (Exception e) {
 			lu.writeException(e);
-			Printer.instance.print(e.getMessage());
+			Printer.instance.print("错误：文件数据读取失败。详细信息：" + e.getMessage());
 		}
 		return null;
 	}
@@ -177,43 +262,48 @@ public class FileBlockUtil {
 	 */
 	public void checkFileBlocks() {
 		Thread checkThread = new Thread(() -> {
-			// 检查是否存在未正确对应文件块的文件节点信息，若有则删除
-			List<Node> nodes = fm.queryAll();
-			for (Node node : nodes) {
-				File block = getFileFromBlocks(node);
-				if (block == null || !block.exists()) {
-					fm.deleteById(node.getFileId());
-				}
-			}
-			// 反向检查是否存在未应文件节点信息的文件块，如有则删除
+			// 检查是否存在未正确对应文件块的文件节点信息，若有则删除，从而确保文件节点信息不出现遗留问题
+			checkNodes("root");
+			// 检查是否存在未正确对应文件节点的文件块，若有则删除，从而确保文件块不出现遗留问题
 			List<File> paths = new ArrayList<>();
 			paths.add(new File(ConfigureReader.instance().getFileBlockPath()));
 			for (ExtendStores es : ConfigureReader.instance().getExtendStores()) {
 				paths.add(es.getPath());
 			}
-			for (File p : paths) {
-				String[] bn = p.list();
-				for (String n : bn) {
-					Node node = fm.queryByPath(n);
-					if (node == null) {
-						File file = null;
-						if (n.startsWith("file_")) {// 存放于主文件系统中
-							// 直接从主文件系统的文件块存放区获得对应的文件块
-							file = new File(ConfigureReader.instance().getFileBlockPath(), n);
-						} else {// 存放于扩展存储区
-							short index = Short.parseShort(n.substring(0, n.indexOf('_')));
-							// 根据编号查到对应的扩展存储区路径，进而获取对应的文件块
-							file = new File(ConfigureReader.instance().getExtendStores().parallelStream()
-									.filter((e) -> e.getIndex() == index).findAny().get().getPath(), n);
-						}
-						if (file != null && file.isFile()) {
-							file.delete();
+			for (File path : paths) {
+				try (DirectoryStream<Path> ds = Files.newDirectoryStream(path.toPath())) {
+					Iterator<Path> blocks = ds.iterator();
+					while (blocks.hasNext()) {
+						File testBlock = blocks.next().toFile();
+						if (testBlock.isFile()) {
+							List<Node> nodes = fm.queryByPath(testBlock.getName());
+							if (nodes == null || nodes.isEmpty()) {
+								testBlock.delete();
+							}
 						}
 					}
+				} catch (IOException e) {
+					Printer.instance.print("警告：文件节点效验时发生意外错误，可能未能正确完成文件节点效验。错误信息：" + e.getMessage());
+					lu.writeException(e);
 				}
 			}
 		});
 		checkThread.start();
+	}
+
+	// 校对文件节点，要求某一节点必须有对应的文件块，否则将其移除（避免出现死节点）
+	private void checkNodes(String fid) {
+		List<Node> nodes = fm.queryByParentFolderId(fid);
+		for (Node node : nodes) {
+			File block = getFileFromBlocks(node);
+			if (block == null) {
+				fm.deleteById(node.getFileId());
+			}
+		}
+		List<Folder> folders = flm.queryByParentId(fid);
+		for (Folder fl : folders) {
+			checkNodes(fl.getFolderId());
+		}
 	}
 
 	/**
@@ -366,4 +456,140 @@ public class FileBlockUtil {
 			}
 		}
 	}
+
+	/**
+	 * 
+	 * <h2>生成指定文件块资源对应的ETag标识</h2>
+	 * <p>
+	 * 该方法用于生产指定文件块的ETag标识，从而方便前端控制缓存。生成规则为：{文件最后修改时间计数}_{文件路径对应的Hash码}。
+	 * </p>
+	 * 
+	 * @author 青阳龙野(kohgylw)
+	 * @param block
+	 *            java.io.File 需要生成的文件块对象，应为文件，但也支持文件夹，或者是null
+	 * @return java.lang.String 生成的ETag值。当传入的block是null或其不存在时，返回空字符串
+	 */
+	public String getETag(File block) {
+		if (block != null && block.exists()) {
+			StringBuffer sb = new StringBuffer();
+			sb.append("\"");
+			sb.append(block.lastModified());
+			sb.append("_");
+			sb.append(block.hashCode());
+			sb.append("\"");
+			return sb.toString().trim();
+		}
+		return "\"0\"";
+	}
+
+	/**
+	 * 
+	 * <h2>插入一个新的文件节点至文件系统数据库中</h2>
+	 * <p>
+	 * 该方法将尝试生成一个新的文件节点并存入文件系统数据库，并确保该文件节点再插入后不会与已有节点产生冲突。
+	 * </p>
+	 * 
+	 * @author 青阳龙野(kohgylw)
+	 * @param fileName
+	 *            java.lang.String 文件名称
+	 * @param account
+	 *            java.lang.String 创建者账户，若传入null则按匿名创建者处理
+	 * @param filePath
+	 *            java.lang.String 文件节点对应的文件块索引
+	 * @param fileSize
+	 *            java.lang.String 文件体积
+	 * @param fileParentFolder
+	 *            java.lang.String 文件的父文件夹ID
+	 * @return kohgylw.kiftd.server.model.Node 操作成功则返回节点对象，否则返回null
+	 */
+	public Node insertNewNode(String fileName, String account, String filePath, String fileSize,
+			String fileParentFolder) {
+		final Node f2 = new Node();
+		f2.setFileId(UUID.randomUUID().toString());
+		if (account != null) {
+			f2.setFileCreator(account);
+		} else {
+			f2.setFileCreator("\u533f\u540d\u7528\u6237");
+		}
+		f2.setFileCreationDate(ServerTimeUtil.accurateToDay());
+		f2.setFileName(fileName);
+		f2.setFileParentFolder(fileParentFolder);
+		f2.setFilePath(filePath);
+		f2.setFileSize(fileSize);
+		int i = 0;
+		// 尽可能避免UUID重复的情况发生，重试10次
+		while (true) {
+			try {
+				if (this.fm.insert(f2) > 0) {
+					if (isValidNode(f2)) {
+						return f2;
+					} else {
+						break;
+					}
+				}
+				break;
+			} catch (Exception e) {
+				f2.setFileId(UUID.randomUUID().toString());
+				i++;
+			}
+			if (i >= 10) {
+				break;
+			}
+		}
+		return null;
+	}
+
+	/**
+	 * 
+	 * <h2>检查指定的文件节点是否存在同名问题</h2>
+	 * <p>
+	 * 该方法用于检查传入节点是否存在冲突问题，一般在新节点插入后执行，若存在冲突会立即删除此节点，最后会返回检查结果。
+	 * </p>
+	 * 
+	 * @author 青阳龙野(kohgylw)
+	 * @param n
+	 *            kohgylw.kiftd.server.model.Node 待检查的节点
+	 * @return boolean 通过检查则返回true，否则返回false并删除此节点
+	 */
+	public boolean isValidNode(Node n) {
+		Node[] repeats = fm.queryByParentFolderId(n.getFileParentFolder()).parallelStream()
+				.filter((e) -> e.getFileName().equals(n.getFileName())).toArray(Node[]::new);
+		if (flm.queryById(n.getFileParentFolder()) == null || repeats.length > 1) {
+			// 如果插入后存在：
+			// 1，该节点没有有效的父级文件夹（死节点）；
+			// 2，与同级的其他节点重名，
+			// 那么它就是一个无效的节点，应将插入操作撤销
+			// 所谓撤销，也就是将该节点的数据立即删除（如果有）
+			fm.deleteById(n.getFileId());
+			return false;// 返回“无效”的判定结果
+		} else {
+			return true;// 否则，该节点有效，返回结果
+		}
+	}
+
+	/**
+	 * 
+	 * <h2>获取一个节点当前的逻辑路径</h2>
+	 * <p>
+	 * 该方法用于获取指定节点当前的完整逻辑路径，型如“/ROOT/doc/test.txt”。
+	 * </p>
+	 * 
+	 * @author 青阳龙野(kohgylw)
+	 * @param n
+	 *            kohgylw.kiftd.server.model.Node 要获取路径的节点
+	 * @return java.lang.String 指定节点的逻辑路径，包含其完整的上级文件夹路径和自身的文件名，各级之间以“/”分割。
+	 */
+	public String getNodePath(Node n) {
+		Folder folder = flm.queryById(n.getFileParentFolder());
+		List<Folder> l = fu.getParentList(folder.getFolderId());
+		StringBuffer pl = new StringBuffer();
+		for (Folder i : l) {
+			pl.append(i.getFolderName() + "/");
+		}
+		pl.append(folder.getFolderName());
+		pl.append("/");
+		pl.append(n.getFileName());
+		return pl.toString();
+	}
+
 }
